#161-added loading text when files are scanned
