+++ src/components/bounce-loader/index.tsx
@@ -1,0 +1,27 @@
+/* eslint-disable max-lines-per-function */
+import { Animated, View } from 'react-native';
+
+import { useBouncingMessage } from '@/core/hooks/use-bouncing-message';
+
+const BounceLoader = ({
+  loadingMessages,
+  className,
+}: {
+  loadingMessages: string[];
+  className?: string;
+}) => {
+  const { fadeValue, loadingMessage } = useBouncingMessage(loadingMessages);
+  return (
+    <View className={className}>
+      {/* Fading Loading Message */}
+      <Animated.Text
+        className="mt-4 font-medium text-white"
+        style={{ opacity: fadeValue }}
+      >
+        {loadingMessage}
+      </Animated.Text>
+    </View>
+  );
+};
+
+export default BounceLoader;
