+++ src/components/end-scroll-placeholder/index.tsx
@@ -29,7 +29,7 @@ export const EndScrollPlaceholder = ({
 
   return (
     <Animated.View
-      className="top-[-100]  items-center justify-center"
+      className=" items-center justify-center"
       style={[{ opacity: fadeAnim }]}
     >
       <TouchableOpacity onPress={onScrollToTop}>
