+++ src/components/image-scanner-modal/index.tsx
@@ -1,11 +1,15 @@
+/* eslint-disable max-lines-per-function */
 import { BlurView } from 'expo-blur';
 import LottieView from 'lottie-react-native';
 import React from 'react';
 import { Image, Modal, View } from 'react-native';
 
+import { LOADING_MESSAGES_IMAGE_SCANNING } from '@/constants/loading-messages';
 import { translate } from '@/core';
-import { Button, Text } from '@/ui';
+import { Button, colors, Text } from '@/ui';
+import { CloseIcon, RetryIcon } from '@/ui/assets/icons';
 
+import BounceLoader from '../bounce-loader';
 import VideoPlayer from '../video';
 import { type IImageScannerModal } from './image-scanner-modal.interface';
 import { imageScannerModalStyles } from './image-scanner-modal.styles';
@@ -54,18 +58,26 @@ const ScanningModal = ({
               </View>
             )}
           </View>
+          {isPending && (
+            <BounceLoader
+              className="mt-10"
+              loadingMessages={LOADING_MESSAGES_IMAGE_SCANNING}
+            />
+          )}
           <View className="mt-10 flex-row justify-center gap-5">
             {!!error && (
               <Button
                 label={translate('general.retry')}
                 onPress={onRetry}
                 disabled={isPending}
+                icon={<RetryIcon color={colors.white} width={18} height={18} />}
               />
             )}
             <Button
               label={translate('general.close')}
               onPress={onClose}
               disabled={isPending}
+              icon={<CloseIcon color={colors.white} width={18} height={18} />}
             />
             {!!error && <Text>{error.toString()}</Text>}
           </View>
