+++ src/constants/loading-messages.ts
@@ -1,0 +1,13 @@
+export const LOADING_MESSAGES_INITIAL_APP_LOAD = [
+  'Loading your personalized experience...',
+  'Synchronizing systems...',
+  'Processing your data...',
+  'Almost ready, just a moment...',
+];
+
+export const LOADING_MESSAGES_IMAGE_SCANNING = [
+  'Scanning your medical imaging, please hold tight...',
+  'Analyzing details, this may take a moment...',
+  'Enhancing accuracy for the best results...',
+  'Almost there, finalizing the scan...',
+];
