+++ src/core/screens/file-preview-screen/index.tsx
@@ -1,6 +1,7 @@
 /* eslint-disable max-lines-per-function */
 import { router, Stack } from 'expo-router';
 import { firebaseAuth } from 'firebase/config';
+import { useColorScheme } from 'nativewind';
 import React, { useState } from 'react';
 import { useTranslation } from 'react-i18next';
 import { ScrollView, View } from 'react-native';
@@ -13,7 +14,6 @@ import CustomHeader from '@/components/cusom-header';
 import ScanningModal from '@/components/image-scanner-modal';
 import ProgressBar from '@/components/progress-bar';
 import PromptSection from '@/components/prompt-section';
-import { useModal } from '@/core/hooks/use-modal';
 import { checkIsVideo } from '@/core/utilities/check-is-video';
 import { getBase64ImageUri } from '@/core/utilities/get-base64-uri';
 import { Button, colors } from '@/ui';
@@ -84,8 +84,8 @@ const FilePreviewScreen = ({
   const [promptMessage, setPromptMessage] = useState('');
   const [isModalVisible, setIsModalVisible] = useState(false);
 
-  const { isVisible: isMediaModalVisible, closeModal, openModal } = useModal();
-
+  const { colorScheme } = useColorScheme();
+  const isDark = colorScheme === 'dark';
   const { mutate: onDecrementScans } = useDecrementScans();
   const {
     i18n: { language },
@@ -168,6 +168,7 @@ const FilePreviewScreen = ({
               className="bg-white pt-20"
               titlePosition="center"
               onGoBack={onGoBack}
+              backIconColor={isDark ? colors.white : colors.black}
             />
           ),
         }}
